fix: make interaction handlers toggleable
