+++ config.json.example
@@ -2,7 +2,8 @@
 "token": "",
 "clientId": "",
 "environment": "",
-"maintenance_mode": false,
+"run_sync_processes": false,
+"handle_interactions": true,
 "DISCORD_VERIFICATION_SECRET": "",
 "FIREBASE_ADMIN_SERVICE_ACCOUNT": {}
 }
