+++ src/index.ts
@@ -1,6 +1,5 @@
 import { Client, Intents } from "discord.js";
-import cron from "node-cron";
-import { maintenance_mode, token } from "../config.json";
+import { handle_interactions, run_sync_processes, token } from "../config.json";
 import db from "./services/admin";
 import { coldUpdateDiscordRolesForUser } from "./utils/coldUpdateDiscordRolesForUser";
 import { connectObserver } from "./utils/connectObserver";
@@ -8,6 +7,7 @@ import { handleNewBlock } from "./utils/handleNewBlock";
 import { handleNFTMoveEvent } from "./utils/handleNFTMoveEvent";
 import { interactionHandler } from "./utils/interactionHandler";
 import { registerCommands } from "./utils/registerCommands";
+import { runSyncProcesses } from "./utils/runSyncProcesses";
 import { updateAllDiscordUserRoles } from "./utils/updateAllDiscordUserRoles";
 import { updateDiscordRolesForUser } from "./utils/updateDiscordRolesForUser";
 
@@ -23,6 +23,7 @@ export class LunarAssistant {
   public handleNewBlock = handleNewBlock;
   public connectObserver = connectObserver;
   public interactionHandler = interactionHandler;
+  public runSyncProcesses = runSyncProcesses;
 
   constructor() {
     // Create a new client instance
@@ -52,56 +53,23 @@ export class LunarAssistant {
     this.client.once("ready", () => {
       // Reregister guild commands for all servers
       this.registerGuildCommands();
+
+      // only add listeners when not in maintenance mode
+      if (!run_sync_processes) {
+        this.runSyncProcesses();
+      }
       // Call the passed onReady function
       onReady();
     });
 
-    // When the bot is added to a server, configure the slash commands
-    this.client.on("guildCreate", registerCommands);
+    if (handle_interactions) {
+      // When the bot is added to a server, configure the slash commands
+      this.client.on("guildCreate", registerCommands);
 
-    // handle slash command interactions
-    this.client.on("interactionCreate", (interaction) =>
-      this.interactionHandler(interaction)
-    );
-
-    // only add listeners when not in maintenance mode
-    if (!maintenance_mode) {
-      // cronjob to update discord roles once a day
-      cron.schedule("0 0 * * *", () => this.updateAllDiscordUserRoles());
-
-      // only start triggering updates after the initial snapshot
-      let initialSnapshot = true;
-
-      // update discord roles whenever a user document changes
-      this.db.collection("users").onSnapshot((querySnapshot) => {
-        if (!initialSnapshot) {
-          const changedDocs = querySnapshot.docChanges();
-          console.log("Docs changed: " + changedDocs.map((doc) => doc.doc.id));
-          changedDocs.reduce(
-            (p, changedDoc) =>
-              p
-                .then(() =>
-                  this.updateDiscordRolesForUser(changedDoc.doc.id).catch(
-                    // ignore errors
-                    (error) => {}
-                  )
-                )
-                .then(
-                  // delay for one second between processing each user
-                  () =>
-                    new Promise((resolve) =>
-                      setTimeout(() => resolve(null), 1000)
-                    )
-                ),
-            new Promise((resolve) => resolve(null))
-          );
-        } else {
-          initialSnapshot = false;
-        }
-      });
-
-      // listen to nft transfer events
-      this.connectObserver();
+      // Handle slash command interactions
+      this.client.on("interactionCreate", (interaction) =>
+        this.interactionHandler(interaction)
+      );
     }
 
     // start the discord bot
